separate smashpy into a separate file, update build files, update pl to >= 2.0
