+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = markermap
-version = 1.0.2
+version = 1.0.3
 author = Wilson Gregory, Nabeel Sarwar
 author_email = wilson.g.gregory@gmail.com
 description = Marker selection library for single-cell RNA-seq data
@@ -24,12 +24,12 @@ install_requires =
     matplotlib
     numpy
     pandas
-    pytorch-lightning
+    pytorch-lightning >= 2.0
     scanpy
     seaborn
-    smashpy >= 0.1.2
     torch
     umap-learn
+    cosg
 
 [options.packages.find]
 where = src
