+++ src/markermap.egg-info/SOURCES.txt
@@ -7,6 +7,7 @@ src/RankCorr/performance.py
 src/RankCorr/rocks.py
 src/markermap/__init__.py
 src/markermap/other_models.py
+src/markermap/smashpy_model.py
 src/markermap/utils.py
 src/markermap/vae_models.py
 src/markermap.egg-info/PKG-INFO
@@ -14,3 +15,6 @@ src/markermap.egg-info/SOURCES.txt
 src/markermap.egg-info/dependency_links.txt
 src/markermap.egg-info/requires.txt
 src/markermap.egg-info/top_level.txt
+tests/test_benchmark.py
+tests/test_dataloaders.py
+tests/test_plots.py
